Change to help B210s, potential overflow fix
I get underflow errors when I use preamble_length=80. The original scapy_radio code had hard-coded a length of 200 preamble symbols. I modified the code to allow up to 200 symbol preambles.

The assertions in the general_work function previously only discounted the padding in the preamble character block. It did not discount the preamble size or the SOF byte. These are added in to attempt to avoid overflows. It should check to see that the data_len is not longer than the longest possible Z-Wave frame+scapyHdr anyway, but it currently does not. TODO?

I modified the std::memcpy parameters to use dereferences of indexes off preamble and temp arrays, rather than performing arithmetic on pointers, which gcc gives warnings about. Now they don't (at least for Z-Wave).